Extract Word Vectors From Current Vector File

This function provides a more convenient wrapper for extract_words. It uses the current vector file, whose location can be found using cbn_get_vectorfile_location and assigned with cbn_set_vectorfile_location.

cbn_extract_word_vectors(words, verbose = FALSE, report_every = 1e+05)

Arguments

words
words to get vectors for

verbose
whether to report on progress

report_every
how often to check in to see if we should stop

Value

a matrix with word vectors as rows